Allow count for commands
Add the ability to enter a count prior to a command to e.g. jump 'count'
times. This is very similar to the behaviour of less.

At the moment this supports relative movements, but not absolute
movements (jumping to a specific line) as ov currently presents an
additional prompt for 'goto line' and doesn't provide a 'goto line' that
defaults to the last line (as less differentiates between `g` and `G`).