Allow reloading file from disk
This reads the currently displayed file from disk which is helpful for
files that change while being displayed in ov.

Reloading a file by default is bound to the `R` key.

Additionally when jumping to the end of the document it will always be
reloaded as well. This is similar to the behaviour of less.

Closes: #32