More less-like behaviour of `g` and `G`
In less `g` and `G` behave the same when prepended with a count.
Without a count `g` jumps to the beginning of the file and `G` jumps to
the end of the file.

This changes conflicts with the current behaviour of ov as both `g` and
`G` are already mapped, but with a different behaviour.

`g` was already a keybinding to jump to a specific line. However,
currently this opens an input mode for the line number (as opposed to
using a count) and doesn't have a default.

`G` was already a keybinding to toggle line numbers. This keybinding was
changed to `o`.